今後の実装のためにインタフェースの導入やテストの書き換えを実施
