--- common/test_data/event_1.json
@@ -1,1 +1,0 @@
-[{"organizers": [{"id": 16963, "name": "Trend Micro"}], "onsite": false, "finish": "2017-06-25T04:00:00+00:00", "description": "Test your knowledge in targeted attacks, Internet of Things (IoT), Industrial Control Systems (ICS), and other cybersecurity issues.\r\n<br>\r\nIn our mission to make the world safe for exchanging digital information, we are launching CTF 2017 Raimund Genes Cup. It\u2019s a cybersecurity competition that combines online qualifier and on-site finals. Top ten teams from the qualifier will enter the finals, hosted in Tokyo in November 2017.\r\n\r\nProgram: Jeopardy Style\r\nDates: June 24 - 25 2017\r\nDuration: 24 hours \r\nStart Time: 13:00 (JST) June 24, 2017\r\nEnd Time: 13:00 (JST) June 25, 2017\r\nLanguage: English\r\nVenue: Online (requires internet connection)\r\nRequirements: Aged over 20", "weight": 0.00, "title": "Trend Micro CTF 2017 - Raimund Genes Cup - Online Qualifier", "url": "https://www.trendmicro.com/tmctf", "is_votable_now": false, "restrictions": "Open", "format": "Jeopardy", "start": "2017-06-24T04:00:00+00:00", "participants": 73, "ctftime_url": "https://ctftime.org/event/475/", "location": "", "live_feed": "https://ctftime.org/live/475/", "public_votable": true, "duration": {"hours": 0, "days": 1}, "logo": "https://ctftime.org/media/events/Profile_Photo_1.jpg", "format_id": 1, "id": 475, "ctf_id": 114}, {"organizers": [{"id": 8186, "name": "nuitduhack"}], "onsite": true, "finish": "2017-06-25T04:00:00+00:00", "description": "As a long-time partner of the Nuit du Hack event, Sysdream has been hosting the Capture The Flag challenge for years. This time, we will welcome the top 10 teams, that passed the online qualification of last April, for a night of competition. The challenge will follow the attack- defense format: every team has an identical network and uses its skills to find, exploit and patch vulnerabilities under constant enemy fire. On site, the public dashboard will display most of the information.", "weight": 0.00, "title": "Nuit du Hack CTF Finals 2017", "url": "https://www.nuitduhack.com/", "is_votable_now": false, "restrictions": "Prequalified", "format": "Attack-Defense", "start": "2017-06-24T19:30:00+00:00", "participants": 3, "ctftime_url": "https://ctftime.org/event/411/", "location": "Paris, France", "live_feed": "", "public_votable": true, "duration": {"hours": 8, "days": 0}, "logo": "https://ctftime.org/media/events/HZV_1.png", "format_id": 2, "id": 411, "ctf_id": 16}, {"organizers": [{"id": 16172, "name": "SharLike"}], "onsite": false, "finish": "2017-06-26T18:00:00+00:00", "description": "III open interregional contest in information security", "weight": 0.00, "title": "AltayCTF-2017", "url": "https://university.altayctf.ru/", "is_votable_now": false, "restrictions": "Open", "format": "Jeopardy", "start": "2017-06-26T10:00:00+00:00", "participants": 23, "ctftime_url": "https://ctftime.org/event/472/", "location": "Russia, Barnaul, Lenina 46", "live_feed": "", "public_votable": false, "duration": {"hours": 8, "days": 0}, "logo": "https://ctftime.org/media/events/DP2qeCghjTU.jpg", "format_id": 1, "id": 472, "ctf_id": 111}]
