+++ const.go
@@ -1,6 +1,7 @@
-package common
+package ctftime
 
 const (
 	BUFSIZE      = 1024
 	API_ENDPOINT = "https://ctftime.org/api/v1"
+	test_dir     = "./test_data/"
 )
